% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_hooks}
\alias{add_hooks}
\title{Add default logging hooks to your pipework project}
\usage{
add_hooks()
}
\description{
Creates a file in your package source R/hooks_.R which
contains code to register pre and post route logging hooks
to your API.
}
