% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_logging}
\alias{add_logging}
\title{Add package code for logging}
\usage{
add_logging()
}
\description{
Creates a file in R/ of your package that
provides functions for setting up and using a logger
with your API.
}
